+++ function/ADC/mega128_ADC.h
@@ -1,2 +1,2 @@
 void ADC_single_initalizer(int adc_num);
-float ADC_single_read(void);
+uint16_t ADC_single_read(void);
